start of script for painting skeleton tracking training mesh
To be follow the approach Microsoft took with training their skeleton
tracking system we need labeled meshes that associate unique colors with
different parts of the body.

The current plan is to build a rendering pipeline based on Blender and
makehuman along with mocap data from CMU for generating our training
data.

This adds the start of a Blender addon script that assumes you have
first selected an armature and a mesh and will then iterate the faces of
the mesh and follows a chosen scheme for painting different parts of the
body.

For now it's just measuring the euclidean distance of vertices from key
bones, with a table look up for colors associated with bones.

Note: The current implementation scales very badly with the complexity
of a mesh because coloring is done on an iteratively increasing distance
threshold (increasing by 5mm each step) and the whole mesh is traversed
for each distance threshold update. E.g. we should skip all faces that
have already been colored in.