blender/paint_rig: skip painted faces
This skips over painted faces sooner for each threshold update
iteration.

It also switches to using the BMesh intermediate representation while I
initially thought it might help with taking a different approach to
reducing the cost of iterations through explicit deletion of faces from
a shadow mesh, but something like this needs more thought.

I have the impression the bmesh structure should be more efficient to
traverse (not checked though) so for now I'll keep the conversion
to/from the bmesh.