blender/rig_paint: support writing out sequential 'greyscale' id labels
Actually we might remove this later if we end up handling the same thing
via a standalone post-processing tool. The issue is that we want to
optimize training costs by doing an offline mapping of the RGB colors to
unique integers that can index into the histograms built while training.