blender/rig_paint: tweaks to tighten up how much we paint joints
One other change for the hands is done to reduce the rig editing that's
necessary in blender to remove the finger joints from the 'Game engine'
rig that's exported by Makehuman. We no longer need to change/resize the
bone positioned at the wrist we can simply delete the fingers and update
the weight painting.