glimpse_data_generator: initial code for adding clothing
This adds an initial 'Add Hoodie' operator that shows how we can add
clothing dynamically to a proportioned body mesh. The clothing will take
weights and colors from a reference clothing object that already exists
in the .blend file.

We use a reference object so we can paint the clothing with the same
operator we paint the body with and we can do this on a standard Tpose
body. This lets us transfer attributes exactly from the reference object
according to the mesh topology so we don't introduce any interpolated
vertex colors.