glimpse_data_generator: support rendering out randomized frames
This adds initial support for iterating through our indexed mocap data
and for each (non blacklisted) file it will retarget the animation on
a set of armatures attached to human meshes with different body
proportions. Each human mesh also has a library of clothes (quite small
currently) that can be selected dynamically. The rendering is done in
batches of 5 frames and then the body and clothing selection are
randomized before continuing.