glimpse_data_generator: improve handling of unreviewed mocap files
The camera is now set up to be able to track one of the spine pose bones
which we update as we switch between different body meshes while
rendering. We now also skip frames where the body is closer than 1.5
meters to avoid frames with the body moving off screen.