glimpse_data_generator: handle save/load of camera fov
We've now switched to rendering with an 83 degree vertical fov which
should be more representative of a mobile phone camera than the previous
49 degrees. The fov is now tracked in the mocap index and when loading
mocap files we take into account older entries without an associated fov
so we can reduce the distance of the camera while increasing the fov so
we get a similar view of the animated pose.

While loading mocap files we differentiate whether we are interactively
reviewing mocap files or doing an unsupervised render of lots of data.
In the former case then if we load a mocap file with no explicit camera
position and orientation saved yet then we keep the existing parameters
assuming they are a reasonable default, but in the non-interactive case
we want the handling to be consistent/deterministic no matter what the
previously loaded file/state was.