Add find-redundant-images.sh script
This script iterates through a directory of rendered images and compares
sequential images with the perceptualdiff tool to determine whether
images are similar and record whether images can be ignored for the
purposes of training our randomized decision forest.

The script only records it's results to stdout so the output should be
saved, and possibly reviewed by hand before deciding to finally delete
the redundant images.