labelling/train.py: Reintroduce threshold loop and add splitOnThreshold
The unrolled threshold loop, while faster than the old loop with
conditions, was still not particularly fast. Replace getLRPixelIndices
with splitOnThreshold which uses the same dynamic_partition method to
split a tensor in two, and use that in both the reintroduced threshold
loop and the pixel collection loop. This both simplifies and speeds up
the code. The root node of a 100-image batch has gone from taking about
6:00 to about 4:40. Still need to investigate how this affects the ability
to scale tree breadth.

Also add some handy node names and the ability to profile the graph,
though the generated Chrome profiles seem to crash Chrome :(