glimpse_data_generator: multiple updates
Includes these changes:

For each mocap we now render with all bodies instead of only switching
the body randomly when switching clothes. This helps increase the number
of output frames.

We no longer consider the camera positions saved in the index.json
database since it's awkward to change the FOV for rendering if we depend
on fixed camera positions. (Notable since we recently updated the FOV in
the .blend file based on the FOV of the Lenovo Phab 2's depth camera.)

We now use the orientation of the body to position the camera.

We avoid frames facing the back of the body.

We randomize the distance and the precise direction of the camera at the
same time we randomize the clothing.

We render output files to a directory named based on the date and time.