json-to-jnt: require either .head or .tail specification
Initially I was hitting an exception because I had json files with more
than just 'head' and 'tail' members associated with each bone. The
script tripped up because it expected it could enumerate all bone
keys and all values would be vectors that could also be averaged if
neither .head or .tail was specified.

I didn't initially grok what the intention of enumerating all keys and
averaging was and in the processes of debugging to fix my issue I've
ended up restricting the flexibility of these map files a little.
Currently I think this is ok since it seemed like a surprising feature
(until I re-read the explanation in the README) that could be used
unwittingly (as I did to start with) and atm I don't know that we need
to be able to do this.

This change added a few defensive checks on what's seen while parsing.