depth2labels: include default palette
Apparently I can't read since I didn't initially see that the tool lets
you specify a filename for another png file to load a palette from but
also think it's convenient if the tool can just output a paletted png
by default without necessarily loading an external palette.

This splits out the code to load a palette for the sake of then using a
copy of write_png_file from image-pre-processor.cc. Later we we can maybe
switch to a shared implementation.

An external palette can be optionally loaded via -p or --palette. A -g
option is added like the image-pre-processing tool for explicitly asking
for greyscale image output.

Some code copy & pasted from image-pre-processing.cc and tweaked for the
coding style. The copied write_png_file opens and closes the FILE
pointer which I think is slightly preferable considering the code wasn't
previously calling fclose() after the write to ensure buffered IO is
flushed before the file descriptor is closed.

Also seems worth noting that to start with I was also having a problem
with the palette in the PNG not being complete (just two entries) and
I couldn't view the data properly in the Gimp for example. I'm not
exactly sure why that was the case and after going in circles for a
while debugging this it started working but I lost track of the before
and after to see where it had really gone wrong.