training/Makefile: be a bit smarter
The previous rules weren't really working well at all because when gcc
is passed multiple source files to compile in one go then it doesn't
necessarily immediately exit for failing to build one of those source
files and can even end up writing a target with an intermediate compile
failure. This meant that a subsequent run of make looks like it succeeds
because the (broken) target looks up to date.

This adds separate %.o rules for building C and C++ code and linker
rules first depend on a set of .o files to be built. These new rules use
the typical toolchain options for spitting out make rules that help
track header dependencies for object files.

The rules now also support a V=1 option for verbose or tidy output.