training: switch everything over to using half.hpp
This avoids us depending on the half float implementation that's
exposed by the openexr api in preparation for sharing code with the
glimpse_android_demo app which we don't want to have depending on
openexr.

For now we use the fully qualified half_float::half type name in places
because we still have some code using openexr.