avoid using = {0} initializers with dynamic sized arrays
This isn't technically allowed by current C standards and clang treats
this as an error. Hopefully gcc's silence implies it did actually
support them as a non-standard feature (as opposed to leaving them
silently un-initialized) but still it seems valuable to be able to build
with clang too for extra static analysis.