pack: update abi in line with python module changes
As I was working on a python module for reading and writing pack files I
git a few issue that are addressed here.

Measuring the size of different property types was broken

There was no getter for blob properties

Putting the total size of the header inside the compressed header is
awkward, due to a chicken egg issue. For new files there's no limit to
the size of the header so we want to compress and then see how much
space is needed.