train_utils: gather files given single data_dir
Assuming a consistent layout for our training data we only need one path
to the top the location of labels, depth, other data should be implicit.

Considering that we are now starting to create top level index files and
top-level meta.json files then there's more reason to be able to work
with a single top-level path otherwise we're going to need to keep
plumbing through new paths for locating new data and risk inconsistent
layouts of data that's cumbersome to share.

Since the location of labels is now inferred then the train_joint_params
tool no longer has a -l option and there's now a boolean -a option to
request that the accuracy of joint inference be checked (which implies
the need to load label images)