loader: add comment and paranoid ABI checks
The fact that __attribute__((vector_size())) results in a type that
naturally aligns according to the total size doesn't seem clearly
documented but at least seems to be the case for gcc. Just adding an
assertion to make sure this always holds true.

Also assert that the RDTHeader is packed without padding between
the first six bytes and the float fov.