Use JSON based joint map
Although the previous CSV format itself was trivial, the data structures
we created were somewhat awkward to juggle.  We had a list of lists
for the labels separate from the array of names, separate from the
n_joints constant and all of this data which constitutes the joint map
was being passed around separately.

We now use json_parse_file() and get a single JSON_Value* pointer
we can pass around for the joint map encapsulating all of the above, and
allows for extending how we describe mappings later if necessary.

The internal data structures created by parson are just arrays so it's
generally reasonably efficient, except for needing a function call to
index into arrays/objects vs pointer indirection of the previous lists.

Considering how the joint map is accessed within the inner loops of
calc_pixel_weights and infer_joints, and also observing that the
indirection of pointers was also far from ideal here this patch
introduces a transient array representation of the label mappings for
those two functions.

The repo now has a training/joint-map.json file which the README.md now
suggests should be used by default for any training.

The gather_train_data code now loads the joint map if given so that it
can determine the number of joints up front and also allocate memory for
joints up front like we do for image data now.