Adds a glimpse_viewer test application
This application work with libfreenect to capture depth images and
pushes that through our label inference + joint inference code and
visualises the results.

This adds a glimpse_context api as the start of a basic C api that could
be used by clients wanting to leverage Glimpse for skeletal tracking.

Note: The glimpse_context is intended to be portable between Android and
Linux and camera device agnostic as much as possible.