Split out device code from viewer
This is a fairly significant refactor, trying to lay some groundwork for
creating a Unity plugin whereby we share some of the Kinect IO code
previously part of glimpse_viewer.

Generally there are three main components to consider:

1) the central plugin or application / middleware (e.g.
   glimpse_viewer.cc)
2) the device code for interacting with cameras (gm_device.cc)
3) the context code for managing tracking (gm_context.cc)

Each component is currently running in a separate thread and to
hopefully help alleviate the pain of safely interacting between the
thread boundaries we are now using events to serialize messages aimed at
the central, plugin component. For now requests from the plugin to
devices or from the plugin to the context are just done with function
calls with careful locking where appropriate.

At a high level control flow goes something like this:

     (4) Events notifying    (1) IO requests for
         tracking results        device frames

    +---------------> +-------------> +---------+
    | Context |       | Plugin |      |  Device |
    +---------+ <--------------+ <--------------+

     (3) Fwd frame to        (2) Events notifying
         start tracking          frames ready

The plugin / middleware is the orchestrator. The context will use events
to notify when it's ready for a new frame that it can use for tracking.
The plugin can request the gm_device api for a frame with requirements
(such must include _DEPTH and/or _LUMINANCE) the gm_device api will then
notify with an event when a frame meeting these requirements is ready.
With a new device frame, this can be forwarded to gm_context which will
start a new iteration of tracking. When tracking completes then the
plugin is notified with an event.

Note: the intention is that the gm_context is not dependent on any of
the gm_device code; the idea being:

- To help keep this component portable
- Allow us to more easily refactor the device code if necessary as we
  work with more types of devices/systems later.

Tripple buffering:

The gm_device component is currently managing back buffers for the depth
and video read from the kinect where a 'back' buffer is kept for IO and
preparation and when ready to be consumed it is swapped and becomes the
'mid' buffer. When the consumer explicitly decides to use the frame then
it will perform another swap where it becomes the 'front' buffer.

Similarly in gm_context then the tracking state is tripple buffered so
new tracking work is done with the 'back' buffer. When tracking
completes then this becomes the 'mid' buffer. When the plugin component
decides to consume the latest tracking state then this becomes the
'front' buffer.