Adds a meson build configuration
As an improvement over our currently hand-rolled Makefile for building
this allows us to build with Meson.

For now the Makefile is kept in parallel, but the plan is to remove it
once we're confident there are no issues / regressions using Meson instead.

Currently the configuration only supports linking against system
provided dependencies, so before we can remove the Makefile we will also
need to support building with the dependencies from
build-third-party.sh, or a Meson subproject based approach.