Adds gm_assets api instead of shim AAssets api
The use of the Android AAssets api is now relegated to an implementation
detail of the gm_assets api on Android. On Android there is one Android
specific entry point gm_android_set_asset_manager() for providing a
global asset manager.

On Linux assets are found relative to the path set in the
GLIMPSE_ASSETS_ROOT environment variable.