Build dependencies via Meson + support cross compilation
This enables building all of our dependencies as Meson subprojects and
cross compiling with the Android NDK.

Some minimal build instructions have been added to the README.md

Note: that due to an upstream bug in Meson it's currently necessary to
build with git master of Meson if building with -Duse_system_libs=false
or cross compiling for Android. This should be fixed in the 0.44.0
release.