Remove deprecated build-third-party.sh
So much of this script was dealing with CMake problems and represents a
frustrating wasted of time. We finally bit the bullet and decided it was
really worth the effort to replace the use of CMake in all of our
dependencies with Meson, which means this ugly script of workarounds is
no longer necessary.