ref count frames and buffers
Instead of the _device code triple buffering the frame buffers we now
maintain three pools of memory for video+depth buffers and the frame
structures and each of these resources is ref counted and recycled via
the pool when their ref drops to zero.

A pool can have a limit on how many buffers can be allocated so that
resource acquisition will eventually become throttled based on the rate
buffers are returned back to the pool if the production rate is too
high. For now the frame pool is limited to five buffers in total (which
indirectly limits the depth/video buffer count too).

This N-buffering approach instead of triple buffering makes it more
practical to support multiple decoupled consumers of frames (such as a
preview/render consumer and a tracking consumer).  Each consumer can
safely expect a frame and its underlying buffers to be immutable so long
as they hold a reference. Previously calls to
gm_device_get_latest_frame() would cause old frames to be recycled and
so all consumers needed to be synchronized with these calls.

This update lets us avoid an immediate copy of frame data within
gm_context_notify_frame() (the copy was required previously because
tracking was not synchronized with gm_device_get_latest_frame() calls as
described above).

This should help with the development of the unity plugin where it's not
convenient to synchronize the rendering with
gm_device_get_latest_frame() calls, if we don't make the render hooks
responsible for glimpse event processing.