Support visualizing heat maps for specific labels
This adds a new 'debug_label' property for the context that lets us pick
a particular label index so we can visualize the probabilities of that
label across frames. The visualization is done as a heat map with the
normalized [0:1] probabilities being mapped to a gradient through:
[black, red, yellow, blue, white]. By default -1 means we want to
visualize the most likely labels.

Note: For now the property shows up as an integer slider which isn't very
convenient if you don't know what specific label indices correspond too.