support enum properties + getters/setters
This allow us to attach enumerants to an integer property that
glimpse_viewer (or later the unity plugin) can introspect and display to
users automatically.

This lets us replace the int slider used for the 'debug_label' property
with a drop down showing human readable label names or 'most likely' as
options.

While updating gm_ui_property this also added support for (optional)
getter/setter functions as I expected these would be required to fix the
kinect properties for setting the tilt/ir brightness. In the end I
didn't use these because I found an alternative way of updating the
properties within the kinect event loop which was preferable for
avoiding spamming the freenect api with lots of redundant calls from the
UI thread. I've left the support for getters/setters since it's likely
they will still be useful later.