Some tidying of the src/python/ module
Sets the license to 'MIT', adds some missing copyright headers and
avoids needing an intermediate libglimpse.so to be built which is
awkward to locate now that we support out-of-tree builds. Since this was
previously the only user of libglimpse.so we no longer build a glimpse
shared library.  README.md updated accordingly.