Move some visualisation-specific work outside of track_skeleton
Previously, track_skeleton did a fair amount of work to support
visualisations, but it did this synchronously. Start moving this
visualisation-specific code into the front-end so we can free up some more
time in the tracking thread and get better tracking frame-rates.

This also removes the optional PCL GroundBasedPeopleDetection code, as it
didn't work well anyway and was just increasing maintenance burden.