ref count tracking state
This lets us avoid lots of tracking state copies while maintaining a
history of tracking state. This also allows us to safely decouple
multiple consumers that can assume tracking state is immutable while a
reference is held. Tracking state allocations are recycled in the same
way as frames and buffers.