log: add gm_throw to use instead of xasprintf for errors
Our error handling is generally dealt with by passing a char **err
pointer when the caller wants to be able to catch a description for
any exceptional errors that might need to be presented to the user.

If err is NULL then this is treated like not catching an exception
and the message is automatically output before aborting the program.

Since xasprintf would only output error messages to stderr then the
message could get lost behind Unity running on Android.

This adds a gm_throw() macro that will redirect the message via our
logging api so that the message can be redirected to the Unity logging
apis (whereby they also show up in the Android adb log).