install_unity_libs: Change rpath to $ORIGIN on install
The dependencies for libglimpse-unity-plugin.so are simply being
installed side-by-side so we can avoid the need to set LD_LIBRARY_PATH
by setting an rpath that will resolve deps in the same directory. Linux
and Android both support the use of '$ORIGIN' within an rpath (although
in the case of Android it should be redundant for us).