devices: tracking running state + add _stop() api
To properly handle the cleanup of resources when destroying a context
then it needs to be possible to stop the capture of device frames and
indirect calls into glimpse_context code before starting to destroy
a context.

Note: we can't just *destroy* the device state before destroying the
context because context destruction will release device resources.