meson: set support_64bit_file_offsets=False in cross-file
The Glimpse branch of Meson that we use to build for Android checks this
property as a way to avoid defining _FILE_OFFSET_BITS=64. Android API
levels < 24 don't support 64 bit file offsets for 32bit targets and
since we want to work with devices like the Lenovo Phab2 which only
supports Android 6.1 we need to avoid requesting 64bit file offset
support.