context: avoid pthread_cond_timedwait
The idea had been to limit the waits to 5 milliseconds before checking
the ctx->destroying state so gm_context_destroy could be sure that it's
possible to reliably stop the tracking thread in a finite amount of
time.

The code was written as if the API takes a relative time to limit the
duration of the wait whereas in fact the API takes an absolute time.

It's not clear how this appeared to work in some cases but in other
cases it resulted in hogging the mutex required to change the
ctx->destroying state or notify of a new frame and so tracking would
effectively hang indefinitely.

Instead of relying on a timeout, gm_context_destroy will now kick
the condition variables that will ensure that any current _wait()
will return to check ctx->destroying.