meson: move java code to src/java
To deal with the restrictions of the meson jar() function this adds a
src/java/meson.build to declare the .jar target which is referenced via
subdir() in our top meson.build.

The install script needed a minor tweak so we don't try and duplicate
the 'src/java/' hierarchy under the install location too.