Adds initial Tango support
Ideally this work could be split up, but it's at least all towards the
same end which is enabling the glimpse_viewer and Unity plugin to work
with data from a Tango camera on Android.

Firstly this adds Tango support to glimpse_device.cc which notably
involved introducing the idea of committing to a device config and
waiting until a DEVICE_READY event is received before starting to
interact with the device. This gives a change for us to connect to the
Tango service asynchronously on Android before we interact with the
Tango api to query things like camera intrinsics.

Since Tango reports a point cloud as opposed to a depth buffer then
for consistency we currently project the cloud into an image which
we should ideally optimize later.

There are some changes to how we configure + build:

The latest version of the glimpse-project/meson branch of meson is
required for building on Android:

  pip3 install --user --upgrade git+https://github.com/glimpse-project/meson

-Dunity_project=/path/to/project should be use instead of --prefix to
locate a Unity project where you want to install the Unity plugin

-Dunity_editor=/unity/editor/prefix should be used when building on
Android so that the .jar package for the Unity player Activity can be
found.

Note that since the .jar package we used to build for Unity was called
GlimpseTest.jar and is now GlimpseUnity.jar then you should make sure to
remove any old GlimpseTest.jar which will conflict when Unity tries to
build an apk.

-Dandroid_tango_sdk=/path/to/unpacked/tango/sdk should be used on
Android to enable the use of Tango

Download the Tango SDK as follows:

  mkdir tango_sdk
  cd tango_sdk
  wget https://developers.google.com/tango/downloads/TangoSDK_Ikariotikos_C.zip
  wget https://developers.google.com/tango/downloads/TangoSupport_Ikariotikos_C.zip
  unzip TangoSDK_Ikariotikos_C.zip
  unzip TangoSupport_Ikariotikos_C.zip
  cp -av */{lib,include} .

An APK can be built for the glimpse_viewer and installed with:

  ninja viewer_apk
  adb install -r glimpse_viewer_apk_stage/glimpse_viewer.apk