Decouple depth and video frame updates
This commit changes how device frame requirements works, making them
requests rather than requirements. gm_device will notify as soon as any
requirement is met and it's up to the client to combine frames if they so
wish.

With this, glimpse_viewer changes to request video frames as quickly as it
can process them, while depth frames will be requested as fast as they
can be tracked. This allows the video output to not be gated by either
depth camera refresh or skeleton tracking speed.