Support rotated cameras
The device layer now has a property that represents the camera's
rotation with respect to the devices current physical orientation so
that we know how to rotate frames before running our label inference
(which is sensitive to frame orientation given how it's trained on
images of upright people).

This rotation is dependent on the physical orientation of the device but
may be different to the rotation of the display depending on the device.

In the case of the Zenfone AR then the Tango camera has a landscape
orientation and while the device reports an un-rotated display in
portrait mode the camera is rotated 90CW relative to the portrait
orientation.

Similar to how display rotations are reported on Android in terms of the
CW rotation applications should apply to render upright, we report the
CW rotation that must be applied to camera frames for them to be
upright. (i.e. 270 degrees for the Zenfone AR in portrait)

There is currently a gm_context_set_camera_rotation that gets called
whenever the camera rotation property changes, but this will probably be
updated soon to instead report the rotation via the device frames
instead.

The tracking objects now include camera intrinsics that have been
rotated according to this property and so the viewer + unity plugin now
refer to these instead of the unrotated ('basis') device intrinsics.

Since the viewer/unity plugin render device video frames directly they
have to take the latest camera rotation into account if they want to
display it upright on screen. For video/depth buffers attached to
tracking objects then the context will have handled rotating these.