Attach camera_rotation to gm_frames
Instead of needing to the rotation on the context we now associate the
rotation with device frames. This simplifies the viewer + unity code
since they are no longer responsible for handling property changed
events and forwarding the state to the context. This also avoids
ambiguity if multiple frames could end up buffered but captured with
different rotations.

This way we should also easily be able to include the rotation when
recording.