Adds an object lifetime breadcrumb trail api
To help with debugging when we've failed to unref resources before
trying to destroy the frame/buffer pools this allows us to track a
history of where objects have moved so that when we can print out a
trail of events for leaked objects. The frames and buffers don't
generally move around a lot or that frequently so there's not really any
concern with the overhead of this extra book keeping for now.

Being a bit lazy:

This commit also squashes together some locking fixes in glimpse_unity
to ensure the render thread won't access invalid frame pointers.

It also squashes a change with how we assert we aren't in the middle of
destroying a context within the render thread hook by have a separate
'stopping' phase that matches how code was recently refactored.