context: allow us to optionally apply depth distortion model
This adds a 'apply_depth_distortion' property so we can optionally apply
the depth camera's distortion model when reprojecting Tango point
clouds.

Notably, even though the Tango header implies that the _POLYNOMIAL_3
model is Brown-Conrady with k1, k2, and k3 it didn't seem obviously
equivalent to the Brown-Conrady model which seems to use even exponents
of 2 4 and 6 when multiplying radial distance with k1, k2 and k3.

Note: we ignore the tangential distortion coefficients currently.

This also adds a 'depth_gap_fill' property so the infilling can be
enabled/disabled at runtime.