unity: reset global 'terminating' state on re-init
When we start terminating the unity plugin state we track this in a
global variable so that the render hook (invoked from unity's render
thread unsynchronized with scripts) will become a NOP and not try and
access the plugin state as its destroyed.

We weren't resetting this when re-initializing the plugin so from the
editor you would only see the video in the background the first that you
play the scene after which the render hook was permanently disabled.