device: make it possible to pause playback
The interface is a little awkward (there isn't just a 'pause' property
or a slider to select a fixed frame) mainly because individual frames in
the recording won't necessarily have depth and video and without both
then tracking won't progress.

This adds a toggle to control whether replay loops or stops on the last
frame. This also adds a slider for changing the last frame. With those
you can effectively stop at an arbitrary point within a recording.

When reaching the 'max_frame' then the io thread will enter into a loop
replaying the last depth and video buffers every 16ms. This way it's
still possible to interactively change debug settings and get immediate
feedback for the frame we're paused on.

Notably this reworks some of the time keeping to ensuring that the
timestamps reported within frames are monotonic over repeated loops of a
recording and continue to increase while paused.