unity: update gm_unity_init() to take a json config string
The GlimpseRuntime class now has a public [Serializable] Config member
where properties can easily be changed via the property editor panel. We
serialize these properties and pass the string to gm_unity_init() to
e.g. provide a convenient way of controlling the device backend that's
used or specify the path to a recording.