Avoid crashing if we can't create a ctx or open a device
This fixes a number of crashes working with the unity plugin when
environment variables like GLIMPSE_ASSETS_ROOT or GLIMPSE_RECORDING_PATH
aren't set and so we fail to create a context or open the device we
expected. With the Unity plugin it's possible to override these
environment variables via properties in the editor so if we see the
errors in the debug console it's not necessary to quit the editor before
trying again.