unity: Avoid overriding GLIMPSE_ASSETS_ROOT
This avoids overriding the GLIMPSE_ASSETS_ROOT environment variable
based on the assetsPath Config property for the unity plugin since
it means we can't ever clear the assetsPath property to revert to
the original GLIMPSE_ASSETS_ROOT value which we've trampled.

This adds a gm_get/set_assets_root() api and avoids having lots of code
assuming that the environment variable is the only way to control this
path.

In the viewer this avoids lots of duplicate logic for defining the path
for recordings (which included checking the assets root path as one of
the fallbacks).