viewer: Add 'real-time AR mode'
Enabled by default on Mobile running with Tango, this avoids showing all
the various debug textures from tracking updates (and avoids the high
cost of uploading those textures) and just lets us switch between the
controls and the video feed.

A skeleton wireframe is drawn on the video feed, similar to the
point-cloud except it uses the timestamp of the video frame for
prediction instead of the depth frame.

This lets us check the base-line performance of our Glimpse tracking on
mobile, whereas the debug view can be bit slow on mobile while it's
presenting a lot of data at once.

The mode can be toggled from the controls view.