Replace plane-removal with motion-based RGBD background subtraction
This commit replaces the ad-hoc plane detection code with a motion-based
background subtraction algorithm based on the paper;

"Background subtraction in people detection framework for RGB-D cameras"

There are two major differences to this paper. Rather than rely on a HOG
to determine confidence of a foreground region being a human, we use our
labelling algorithm and then the resulting confidence and skeleton
distance. This isn't always reliable, however, and we also expect to be
able to track a scene that starts off with the human already in shot, so
we have a secondary, naive human detection process based on a focal
point, basic clustering and naive floor removal.

When we detect a human after a period of no human detection, we clear the
background subtraction codebook of any codewords that may correspond to
foreground objects. This way, if the codebook was corrupted by a failed
tracking of a human subject, it hopefully won't affect the tracking of
subsequent frames where we rely on the motion detection.

The second major difference is that we also expect to be able to work
with a moving camera. So using the pose data of the camera, we establish
a start position and we transform all subsequent frames to the start
position before doing depth classification. When the pose changes
substantially, we reset the start position and clear the depth
classification data.