ios: enable AVFoundation camera support in viewer
This adds a new glimpse_device backend that adds support for capturing
video and dual-camera depth via the AVFoundation API on iOS.

Since AVFoundation reports camera intrinsics as a part of the
video/depth frames and not constant for a given device then this patch
also conflates updating the glimpse_device API to forward intrinsics as
part of a gm_frame and to not allow querying the device's intrinsics.

Recordings now save intrinsics with each frame but we maintain support
for loading recordings with just one set of intrinsics for the entire
recording.