viewer: repeat joint recordings after device open
If we want to record joints to json we can set
GLIMPSE_RECORD_N_JOINT_FRAMES but this interfaces is awkward if you want
to first select a recording to open in the viewer. Now if a new device
is opened e.g. due to selecting a new recording then any requested joint
recording will re-start.