WIP: Add a simpler human-detector to help motion-based bg subtraction
An issue with motion-based background subtraction for our use-case is that
the subject is often within frame at the start of tracking. With the
chosen algorithm, if at any time points that are part of a human end up
not being classified as human, it badly affects the quality of the depth
classification.

To help with this, this commit adds a simpler focus-point based human
finder, and only does background subtraction on frames that are
successfully tracking a human.