Add an extra classification flag for visualisation and clear less
Add an extra 'CAN' (short for 'candidate') flag in the depth classification
that we can use to visualise when a human cloud was picked out, but
failed to track due to confidence/distance criteria.

To make that flag more useful, clear the depth segmentation data less. Now
it's cleared only when a human is newly tracked, not when humans stop
tracking, so we can better see when/how things fail.