Create debug point cloud in glimpse_context.cc
By making glimpse_context.cc more responsible for describing the point
cloud instead of glimpse_viewer then it should become more practical to
expose more internal tracking state via this view.

To start with this lets us select a number of points in the tracking
pipeline to capture a point cloud and we can texture the cloud with the
video or rgb depth values.

This also makes it easy to draw coloured lines in the same coordinate
space as the cloud for visualization.

Unlike other debug views which are pulled explicitly by the viewer and
created lazily this data needs to be calculated during tracking. When
the viewer is in realtime AR mode then it will configure the context to
not create any debug point cloud data to avoid the overhead.