context: fix + comment surprising special case
On tango then at the same time as projecting a point cloud into a depth
(rotated) buffer in copy_and_rotate_depth_buffer we also initialize a
pcl point cloud so we don't have to iterate over the data twice. We also
do min/max thresholding and check for NAN values.

This is a somewhat surprising special case optimisation for one format
in copy_and_rotate_depth_buffer and gm_context_track_skeleton has a
corresponding check so it knows that for all other cases it still needs
to initialize a pcl point cloud and handle min/max thresholding.

Notably we only hit the special case in copy_and_rotate_depth_buffer if
ctx->apply_depth_distortion == false but we also only hit the
corresponding code in gm_context_track_skeleton if
apply_depth_distortion was == false when it should have been handling
the opposite / all other cases.

This fixes the corresponding check in gm_context_track_skeleton and also
adds a number of comments in both functions so it's clear that both
pieces of code need to remain in sync, until we handle all formats
consistently.